New add_theme_support HTML5 Feature
Thematic is changing to allow child themes to control what markup it
uses, this was previously something that was set in the WP Admin, this
way is much better. This feature isn't pulled into the master
development branch yet, but just getting ready for it since just falls
back to doing nothing.

Also, a few minor formatting revisions.